reworked some display methods, some progress on editor, and more
- I had to figure a better way to update text by sections, so I fixed a few things. This makes it more
efficient in the editor.

- WRT the editor, I decided to use bidirectional chained list for words and paragraphs, which allow more
efficient word replacement, removal, or addition. The editor at this point is definitely the most complex part
of the task, lots of things to plan for, and that includes how to display things at the proper position on
screen too. I might end splitting things between text processing and text display...

- I also added the shutdown/reset functions, and corresponding screens. I need to figure how to do sleep,
but that will likely require an external switch for powering on/off: the pi always stays powered currently, if
the USB-C port is connected to power. I use a USB cord with a switch on it (Canakit Pi4 kit) but it's temporary, I
don't want to use a hard switch eventually.